backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "numpy>=2.2.6",
    "openai>=1.84.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pyglet<2.0",
    "pyrender>=0.1.45",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "trimesh>=4.6.10",
    "uvicorn>=0.34.3",
]
